backend/requirements.txt
Django
djangorestframework
python-jose
requests